13. Signout Api & Admin Dashboard

+++ frontend/src/actions/auth.actions.js
@@ -27,30 +27,6 @@ export const login = (user) => {
   };
 };
 
-// export const signup = (user) => {
-//   return async (dispatch) => {
-//     const res = await axiosIntance.post(`/admin/signup`, { ...user });
-//     dispatch({ type: authConstants.LOGIN_REQUEST });
-//     if (res.status === 201) {
-//       const { message } = res.data;
-//       dispatch({
-//         type: authConstants.LOGIN_SUCCESS,
-//         payload: {
-//           token,
-//           user,
-//         },
-//       });
-//     } else {
-//       if (res.status === 400) {
-//         dispatch({
-//           type: authConstants.LOGIN_FAILURE,
-//           payload: { error: res.data.error },
-//         });
-//       }
-//     }
-//   };
-// };
-
 export const isUserLoggedIn = () => {
   return async (dispatch) => {
     const token = localStorage.getItem("token");
@@ -74,7 +50,16 @@ export const isUserLoggedIn = () => {
 
 export const signout = () => {
   return async (dispatch) => {
-    localStorage.clear();
     dispatch({ type: authConstants.LOGOUT_REQUEST });
+    const res = await axiosIntance.post(`/admin/signout`);
+    if (res.status === 200) {
+      localStorage.clear();
+      dispatch({ type: authConstants.LOGOUT_SUCCESS });
+    } else {
+      dispatch({
+        type: authConstants.LOGOUT_FAILURE,
+        payload: { error: res.data.error },
+      });
+    }
   };
 };
